+++ deegree-datastores/deegree-tilestores/deegree-tilestore-remotewmts/src/main/java/org/deegree/tile/persistence/remotewmts/RemoteWMTSTileDataLevel.java
@@ -109,7 +109,7 @@ class RemoteWMTSTileDataLevel implements TileDataLevel {
     }
 
     @Override
-    public Tile getTile( int x, int y ) {
+    public Tile getTile( long x, long y ) {
         if ( matrix.getNumTilesX() <= x || matrix.getNumTilesY() <= y || x < 0 || y < 0 ) {
             return null;
         }
